Organise storage calls and initialisation
