+++ src/mappings/asset-tx-payment/decode.ts
@@ -1,8 +1,6 @@
 import * as ss58 from '@subsquid/ss58';
-import { AssetMetadata as AssetMetadata_v48 } from '../../types/v48';
-import { AssetMetadata as AssetMetadata_v49, Asset_ForeignAsset } from '../../types/v49';
-import { events, storage } from '../../types';
-import { Block, Event } from '../../processor';
+import { events } from '../../types';
+import { Event } from '../../processor';
 
 export const decodeAssetTxFeePaidEvent = (event: Event): AssetTxFeePaidEvent => {
   let decoded: { who: string; actualFee: bigint; assetId?: any };
@@ -19,23 +17,6 @@ export const decodeAssetTxFeePaidEvent = (event: Event): AssetTxFeePaidEvent => 
     actualFee: BigInt(decoded.actualFee),
     assetIdValue: Number(decoded.assetId),
   };
-};
-
-export const decodeMetadataStorage = async (
-  block: Block,
-  assetId: Asset_ForeignAsset
-): Promise<AssetMetadata_v48 | AssetMetadata_v49 | undefined> => {
-  let assetMetadata: AssetMetadata_v48 | AssetMetadata_v49 | undefined;
-  if (storage.assetRegistry.metadata.v42.is(block)) {
-    assetMetadata = await storage.assetRegistry.metadata.v42.get(block, assetId);
-  } else if (storage.assetRegistry.metadata.v48.is(block)) {
-    assetMetadata = await storage.assetRegistry.metadata.v48.get(block, assetId);
-  } else if (storage.assetRegistry.metadata.v49.is(block)) {
-    assetMetadata = await storage.assetRegistry.metadata.v49.get(block, assetId);
-  } else if (storage.assetRegistry.metadata.v51.is(block)) {
-    assetMetadata = await storage.assetRegistry.metadata.v51.get(block, assetId);
-  }
-  return assetMetadata;
 };
 
 interface AssetTxFeePaidEvent {
