+++ src/mappings/asset-tx-payment/index.ts
@@ -2,7 +2,8 @@ import { HistoricalAccountBalance } from '../../model';
 import { Asset_ForeignAsset } from '../../types/v49';
 import { TREASURY_ACCOUNT, _Asset, extrinsicFromEvent, formatAssetId } from '../../helper';
 import { Block, Event } from '../../processor';
-import { decodeAssetTxFeePaidEvent, decodeMetadataStorage } from './decode';
+import { decodeMetadataStorage } from '../asset-registry/decode';
+import { decodeAssetTxFeePaidEvent } from './decode';
 
 export const assetTxFeePaid = async (block: Block, event: Event): Promise<HistoricalAccountBalance[] | undefined> => {
   const { accountId, actualFee, assetIdValue } = decodeAssetTxFeePaidEvent(event);
